replace c style casts with c++ casts

+++ source/video/VideoView.cpp
@@ -40,8 +40,8 @@ void VideoView::fitToWindow()
 {
 	int width = this->width();
 	int height = this->height();
-	float imgRatio = (float)_displayImage.cols/(float)_displayImage.rows;
-	float windowRatio = (float)width/(float)height;
+	float imgRatio = static_cast<float>(_displayImage.cols) / _displayImage.rows;
+	float windowRatio = static_cast<float>(width) / height;
 	if(windowRatio < imgRatio) 
 	{
 		_zoomFactor = _displayImage.rows/(width/imgRatio);
@@ -157,7 +157,7 @@ void VideoView::paintGL()
 	//set pixel storage mode to byte alignment
 	glPixelStorei(GL_UNPACK_ALIGNMENT, 1 );
 	//and define number of pixels in a row
-	glPixelStorei(GL_UNPACK_ROW_LENGTH, (int) imageCopy.step/imageCopy.channels());
+	glPixelStorei(GL_UNPACK_ROW_LENGTH, static_cast<int>(imageCopy.step)/imageCopy.channels());
 	cv::Mat tile = imageCopy(cv::Range(r, height),
 		cv::Range(c, width));
 	glTexSubImage2D(GL_TEXTURE_2D, 0, c, r, tile.cols, tile.rows, GL_BGR, GL_UNSIGNED_BYTE, tile.ptr());
@@ -212,8 +212,8 @@ QPoint VideoView::unprojectMousePos(QPoint mouseCoord)
 	glGetDoublev( GL_PROJECTION_MATRIX, projection );
 	glGetIntegerv( GL_VIEWPORT, viewport );
 	GLint isOnPicture = gluUnProject(mouseCoord.x(), viewport[3] - mouseCoord.y(), 0, modelview, projection, viewport, &posX, &posY, &posZ);
-	pictureCoord.setX((int)posX);
-	pictureCoord.setY((int)posY);
+	pictureCoord.setX(static_cast<int>(posX));
+	pictureCoord.setY(static_cast<int>(posY));
 	return pictureCoord;
 }
